using From to transform from generic whiteflag message to specific and then trait for encode form compatibility
