+++ src/wf_models/mod.rs
@@ -5,17 +5,11 @@ use std::collections::HashMap;
 #[cfg(test)]
 mod test;
 
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all(deserialize = "camelCase"))]
-pub struct MessageHeader {
-    pub prefix: String,
-    pub version: String,
-    pub encryption_indicator: String,
-    pub duress_indicator: String,
-    pub message_code: char,
-    pub reference_indicator: String,
-    pub referenced_message: String,
-}
+mod authentication;
+mod header;
+
+pub use authentication::AuthenticationMessage;
+pub use header::MessageHeader;
 
 #[derive(Serialize, Deserialize)]
 #[serde(rename_all(deserialize = "camelCase"))]
@@ -26,44 +20,26 @@ pub struct WhiteflagMessage {
     pub body: HashMap<String, Value>,
 }
 
-#[derive(Serialize, Deserialize)]
-#[serde(rename_all(deserialize = "camelCase"))]
-pub struct AuthenticationMessage {
-    #[serde(flatten)]
-    pub header: MessageHeader,
-    pub verification_method: String,
-    pub verification_data: String,
-}
-
-impl MessageHeader {
-    pub fn to_field_values(self) -> Vec<String> {
-        vec![
-            self.prefix,
-            self.version,
-            self.encryption_indicator,
-            self.duress_indicator,
-            self.message_code.to_string(),
-            self.reference_indicator,
-            self.referenced_message,
-        ]
-    }
-}
-
-impl AuthenticationMessage {
-    pub fn to_field_values(self) -> Vec<String> {
-        [
-            self.header.to_field_values(),
-            vec![self.verification_method, self.verification_data],
-        ]
-        .concat()
-    }
+pub trait WhiteflagEncodeCompatible {
+    fn to_field_values(self) -> Vec<String>;
 }
 
 impl WhiteflagMessage {
-    pub fn to_wf_message(&self) {
+    pub fn from_json<T: AsRef<str>>(json: T) -> WhiteflagMessage {
+        let wf_message: WhiteflagMessage = serde_json::from_str(json.as_ref()).unwrap();
+        wf_message
+    }
+}
+
+impl WhiteflagEncodeCompatible for WhiteflagMessage {
+    fn to_field_values(self) -> Vec<String> {
         match &self.header.message_code {
-            'A' => {}
+            'A' => {
+                let auth: AuthenticationMessage = self.into();
+            }
             _ => {}
-        }
+        };
+
+        vec![]
     }
 }
